# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Clipaso API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "type": "array",
                  "title": "Response List Jobs Jobs Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Delete Job",
        "operationId": "delete_job_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/more": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "More Clips",
        "description": "Busca más clips en el mismo vídeo (sin volver a subirlo ni gastar minutos).",
        "operationId": "more_clips_jobs__job_id__more_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoreClipsIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/archive": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Archive Link",
        "description": "Enlace temporal para descargar el ZIP (el navegador no puede enviar la cabecera de sesión en una descarga).",
        "operationId": "archive_link_jobs__job_id__archive_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArchiveOut"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArchiveOut": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "ArchiveOut"
      },
...
      "CaptionStyle": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "font": {
            "type": "string",
            "enum": [
              "Archivo Black",
              "Anton",
              "Bebas Neue",
              "Poppins",
              "Luckiest Guy",
              "Montserrat",
              "Oswald",
              "Bangers",
              "Rubik",
              "Permanent Marker",
              "Inter"
            ],
            "title": "Font",
            "default": "Archivo Black"
          },
          "text_color": {
            "type": "string",
            "pattern": "^[0-9A-Fa-f]{6}$",
            "title": "Text Color",
            "default": "FFFFFF"
          },
          "highlight_color": {
            "type": "string",
            "pattern": "^[0-9A-Fa-f]{6}$",
            "title": "Highlight Color",
            "default": "00E5FF"
          },
          "scale": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 50.0,
            "title": "Scale",
            "description": "Tamaño del texto, en % del tamaño base del formato.",
            "default": 100
          },
          "position": {
            "type": "string",
            "enum": [
              "bottom",
              "middle",
              "top"
            ],
            "title": "Position",
            "default": "bottom"
          },
          "y": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 95.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Y",
            "description": "Altura exacta (% desde arriba); si falta, `position`."
          },
          "uppercase": {
            "type": "boolean",
            "title": "Uppercase",
            "default": true
          },
          "outline": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Outline",
            "description": "Grosor del contorno (0 = sin contorno).",
            "default": 4
          },
          "outline_color": {
            "type": "string",
            "pattern": "^[0-9A-Fa-f]{6}$",
            "title": "Outline Color",
            "default": "000000"
          },
          "shadow": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Shadow",
            "description": "Distancia de la sombra (0 = sin sombra).",
            "default": 2
          },
          "box": {
            "type": "boolean",
            "title": "Box",
            "default": false
          },
          "box_color": {
            "type": "string",
            "pattern": "^[0-9A-Fa-f]{6}$",
            "title": "Box Color",
            "default": "000000"
          },
          "box_opacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Box Opacity",
            "default": 100
          },
          "animation": {
            "type": "string",
            "enum": [
              "highlight",
              "pop",
              "karaoke",
              "appear",
              "none"
            ],
            "title": "Animation",
            "description": "Cómo se marca la palabra que se está diciendo.",
            "default": "highlight"
          },
          "max_words": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 1.0,
            "title": "Max Words",
            "description": "Palabras como máximo en pantalla a la vez.",
            "default": 3
          }
        },
        "type": "object",
        "title": "CaptionStyle",
        "description": "Estilo de subtítulos tal como lo elige el usuario."
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreateIn": {
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "clips",
              "subtitle"
            ],
            "title": "Mode",
            "description": "clips: la IA elige los mejores momentos · subtitle: el vídeo entero con subtítulos.",
            "default": "clips"
          },
          "subtitle_language": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z]{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle Language",
            "description": "Solo subtitular: traducir los subtítulos a este idioma."
          },
          "format": {
            "type": "string",
            "enum": [
              "vertical",
              "square",
              "horizontal",
              "original"
            ],
            "title": "Format",
            "default": "vertical"
          },
          "duration": {
            "type": "string",
            "enum": [
              "auto",
              "short",
              "medium",
              "long"
            ],
            "title": "Duration",
            "default": "auto"
          },
          "topic": {
            "type": "string",
            "maxLength": 200,
            "title": "Topic",
            "description": "Tema opcional: «momentos donde hablo de dinero».",
            "default": ""
          },
          "keep_source": {
            "type": "boolean",
            "title": "Keep Source",
            "description": "Conservar el original para editar clips y pedir más.",
            "default": true
          },
          "branding": {
            "type": "boolean",
            "title": "Branding",
            "description": "Añadir la marca personal del usuario (si la tiene).",
            "default": true
          },
          "caption_style": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaptionStyle"
              },
              {
                "type": "null"
              }
            ],
            "description": "Si falta, el estilo por defecto del usuario."
          },
          "upload_id": {
            "type": "string",
            "title": "Upload Id"
          },
          "max_clips": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Max Clips",
            "default": 3
          },
          "language": {
            "type": "string",
            "pattern": "^(auto|[a-z]{2})$",
            "title": "Language",
            "default": "es"
          },
          "trim_start": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim Start",
            "description": "Usar solo un tramo del vídeo (si no se recortó ya en el navegador): inicio en segundos."
          },
          "trim_end": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim End"
          }
        },
        "type": "object",
        "required": [
          "upload_id"
        ],
        "title": "JobCreateIn"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "done",
              "failed",
              "expired"
            ],
            "title": "Status"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "video_minutes": {
            "type": "number",
            "title": "Video Minutes"
          },
          "clip_count": {
            "type": "integer",
            "title": "Clip Count"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "options": {
            "$ref": "#/components/schemas/JobOptions"
          },
          "frame": {
            "type": "string",
            "enum": [
              "vertical",
              "square",
              "horizontal"
            ],
            "title": "Frame",
            "description": "Encuadre de los clips (en el formato «original», el del vídeo)."
          },
          "can_edit": {
            "type": "boolean",
            "title": "Can Edit",
            "description": "Se conserva el original: se puede editar y pedir más clips."
          },
          "more_clips_available": {
            "type": "integer",
            "title": "More Clips Available"
          },
          "more_clips_task": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TaskOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "clips": {
            "items": {
              "$ref": "#/components/schemas/ClipOut"
            },
            "type": "array",
            "title": "Clips"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "stage",
          "progress",
          "video_minutes",
          "clip_count",
          "thumbnail_url",
          "created_at",
          "finished_at",
          "expires_at",
          "error_code",
          "error_message",
          "options",
          "frame",
          "can_edit",
          "more_clips_available",
          "more_clips_task",
          "clips"
        ],
        "title": "JobOut"
      },
      "JobSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "done",
              "failed",
              "expired"
            ],
            "title": "Status"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "video_minutes": {
            "type": "number",
            "title": "Video Minutes"
          },
          "clip_count": {
            "type": "integer",
            "title": "Clip Count"
          },
          "thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status",
          "stage",
          "progress",
          "video_minutes",
          "clip_count",
          "thumbnail_url",
          "created_at",
          "finished_at",
          "expires_at"
        ],
        "title": "JobSummary"
      },
...
      "MoreClipsIn": {
        "properties": {
          "count": {
            "type": "integer",
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Count",
            "default": 3
          },
          "topic": {
            "type": "string",
            "maxLength": 200,
            "title": "Topic",
            "default": ""
          }
        },
        "type": "object",
        "title": "MoreClipsIn"
      },
...
    }
  }
}